requirements.txt
asgiref==3.2.7
Django==3.0.5
django-enumfield==2.0.1
django-environ==0.4.5
djangorestframework==3.11.0
environ==1.0
Pillow==7.1.2
psycopg2==2.8.5
python-dotenv==0.13.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
